requirements.txt
numpy>=1.21.0,<2.0.0
timm==0.6.12
wandb
pre-commit>=2.15.0,<3.0.0
black[jupyter]>=21.8,<23
flake8>=3.8,<5
isort[colors]>=5.8,<6
pydocstyle[toml]>=6.1,<7
pyupgrade>=1.24,<3
pytest>=6.1.2,<8
pytest-cov>=2.4,<4
types-requests>=2.28.9,<2.30.0
types-python-dateutil>=2.8.19,<2.9.0
kornia==0.6.9
torchgeo==0.5.2
classy-vision==0.6.0
tensorboard
opencv-python
